SED-1054 refactor functions table, reuse datasource definition
